feat: [PIE-3749]: Fix cyclic dependency
