+++ src/modules/10-common/components/ConfigureOptions/AllowedValuesField.tsx
@@ -11,9 +11,9 @@ import { FormikContextType, yupToFormErrors } from 'formik'
 import { Button, ButtonVariation, FormInput, Layout, MultiSelectOption, SelectOption } from '@harness/uicore'
 import { Position } from '@blueprintjs/core'
 import { useStrings, String } from 'framework/strings'
+import type { StringKeys } from 'framework/strings/StringsContext'
 import { ALLOWED_VALUES_TYPE } from './constants'
 import { VALIDATORS } from './validators'
-import type { StringKeys } from 'framework/strings/StringsContext'
 import type { FormValues } from './ConfigureOptionsUtils'
 
 import css from './ConfigureOptions.module.scss'
