+++ src/modules/10-common/components/MultiTypeDuration/MultiTypeDuration.tsx
@@ -24,15 +24,11 @@ import {
   FormikTooltipContext
 } from '@wings-software/uicore'
 import { get } from 'lodash-es'
-import * as Yup from 'yup'
 import { useStrings } from 'framework/strings'
 import { ConfigureOptions, ConfigureOptionsProps } from '@common/components/ConfigureOptions/ConfigureOptions'
 import { errorCheck } from '@common/utils/formikHelpers'
+import { getDurationValidationSchema, GetDurationValidationSchemaProps, isValidTimeString } from './helper'
 import css from './MultiTypeDuration.module.scss'
-
-export function isValidTimeString(value: string): boolean {
-  return !DurationInputHelpers.UNIT_LESS_REGEX.test(value) && DurationInputHelpers.VALID_SYNTAX_REGEX.test(value)
-}
 
 function MultiTypeDurationFixedTypeComponent(
   props: FixedTypeComponentProps & MultiTypeDurationProps['inputGroupProps']
@@ -209,67 +205,7 @@ export function FormMultiTypeDuration(props: FormMultiTypeDurationProps): React.
 
 export const FormMultiTypeDurationField = connect(FormMultiTypeDuration)
 
-export interface GetDurationValidationSchemaProps {
-  minimum?: string
-  maximum?: string
-  inValidSyntaxMessage?: string
-  minimumErrorMessage?: string
-  maximumErrorMessage?: string
-}
-
-export function getDurationValidationSchema(
-  props: GetDurationValidationSchemaProps = {}
-): Yup.StringSchema<string | undefined> {
-  const { minimum = '1s', maximum = '53w' } = props
-
-  if (typeof minimum === 'string' && !isValidTimeString(minimum)) {
-    throw new Error(`Invalid format "${minimum}" provided for minimum value`)
-  }
-
-  if (typeof maximum === 'string' && !isValidTimeString(maximum)) {
-    throw new Error(`Invalid format "${maximum}" provided for maximum value`)
-  }
-
-  return Yup.string().test({
-    test(value: string): boolean | Yup.ValidationError {
-      const { inValidSyntaxMessage, maximumErrorMessage, minimumErrorMessage } = props
-
-      if (!value) return true
-
-      if (getMultiTypeFromValue(value) !== MultiTypeInputType.FIXED) {
-        return true
-      }
-
-      if (typeof value === 'string' && !isValidTimeString(value)) {
-        return this.createError({ message: inValidSyntaxMessage || 'Invalid syntax provided' })
-      }
-
-      if (typeof minimum === 'string') {
-        const minTime = parseStringToTime(minimum)
-        const time = parseStringToTime(value)
-
-        if (time < minTime) {
-          return this.createError({
-            message: minimumErrorMessage || `Value must be greater than or equal to "${timeToDisplayText(minTime)}"`
-          })
-        }
-      }
-
-      if (typeof maximum === 'string') {
-        const maxTime = parseStringToTime(maximum)
-        const time = parseStringToTime(value)
-
-        if (time > maxTime) {
-          return this.createError({
-            message: maximumErrorMessage || `Value must be less than or equal to "${timeToDisplayText(maxTime)}"`
-          })
-        }
-      }
-
-      return true
-    }
-  })
-}
+export { getDurationValidationSchema, GetDurationValidationSchemaProps, isValidTimeString }
 
 export interface DurationInputForInputSetProps extends Omit<IFormGroupProps, 'label' | 'placeholder'> {
   onChange?(str: string): void
